Move MainchainEpochConfig and epoch derivation logic to domain crate
