+++ mainchain-follower/db-sync-follower/src/block/tests.rs
@@ -1,8 +1,8 @@
 use crate::block::{BlockDataSource, BlockDataSourceImpl, BlocksCache};
 use chrono::{NaiveDateTime, TimeDelta};
-use epoch_derivation::{Duration, MainchainEpochConfig, Timestamp};
 use hex_literal::hex;
 use main_chain_follower_api::block::MainchainBlock;
+use sidechain_domain::mainchain_epoch::{Duration, MainchainEpochConfig, Timestamp};
 use sidechain_domain::{McBlockHash, McBlockNumber, McEpochNumber, McSlotNumber};
 use sqlx::PgPool;
 use std::str::FromStr;
