+++ node/src/tests/mock.rs
@@ -4,10 +4,10 @@ use crate::inherent_data::CreateInherentDataConfig;
 use crate::main_chain_follower::DataSources;
 use crate::tests::runtime_api_mock::TestApi;
 use chain_params::SidechainParams;
-use epoch_derivation::MainchainEpochConfig;
 use hex_literal::hex;
 use main_chain_follower_api::mock_services::TestDataSources;
 use sc_consensus_aura::SlotDuration;
+use sidechain_domain::mainchain_epoch::MainchainEpochConfig;
 use sidechain_domain::*;
 use sidechain_slots::{ScSlotConfig, SlotsPerEpoch};
 use sp_core::offchain::{Duration, Timestamp};
