+++ partner-chains-cli/Cargo.toml
@@ -27,7 +27,6 @@ cli-commands = { workspace = true }
 chain-params = { workspace = true, features = ["std"] }
 thiserror = { workspace = true }
 tempfile = { workspace = true }
-epoch-derivation = { workspace = true, features = ["std"] }
 secp256k1 = { workspace = true, features = ["std"] }
 colored = { version = "2.1.0" }
