+++ primitives/domain/Cargo.toml
@@ -21,9 +21,13 @@ num-derive = { workspace = true }
 sp-std = { workspace = true }
 lazy_static = { workspace = true }
 blake2b_simd = { workspace = true }
+figment = { workspace = true, optional = true }
+thiserror = { workspace = true, optional = true }
+sp-runtime = { workspace = true, optional = true }
 
 [dev-dependencies]
 serde_json = { workspace = true }
+frame-support = { workspace = true }
 
 [features]
 default = ["std"]
@@ -33,6 +37,10 @@ std = [
 	"parity-scale-codec/std",
 	"scale-info/std",
 	"num-bigint/std",
+	"figment",
+	"thiserror",
+	"sp-runtime",
+	"sp-runtime/std"
 ]
 serde = [
 	"dep:serde",
