+++ primitives/domain/src/mainchain_epoch.rs
@@ -1,7 +1,7 @@
+use crate::{McEpochNumber, McSlotNumber};
 #[cfg(feature = "std")]
 use parity_scale_codec::Decode;
 use parity_scale_codec::Encode;
-use sidechain_domain::{McEpochNumber, McSlotNumber};
 pub use sp_core::offchain::{Duration, Timestamp};
 
 #[derive(Debug, Clone, PartialEq)]
@@ -133,6 +133,30 @@ impl MainchainEpochDerivation for MainchainEpochConfig {
 mod tests {
 	use super::*;
 	use frame_support::{assert_err, assert_ok};
+
+	#[test]
+	fn read_epoch_config_from_env() {
+		figment::Jail::expect_with(|jail| {
+			set_mainchain_env(jail);
+			assert_ok!(
+				MainchainEpochConfig::read_from_env(),
+				MainchainEpochConfig {
+					first_epoch_timestamp_millis: Timestamp::from_unix_millis(10),
+					first_epoch_number: 100,
+					epoch_duration_millis: Duration::from_millis(1000),
+					first_slot_number: 42,
+				}
+			);
+			Ok(())
+		});
+	}
+
+	fn set_mainchain_env(jail: &mut figment::Jail) {
+		jail.set_env("MC__FIRST_EPOCH_TIMESTAMP_MILLIS", 10);
+		jail.set_env("MC__FIRST_EPOCH_NUMBER", 100);
+		jail.set_env("MC__EPOCH_DURATION_MILLIS", 1000);
+		jail.set_env("MC__FIRST_SLOT_NUMBER", 42);
+	}
 
 	fn test_mc_epoch_config() -> MainchainEpochConfig {
 		MainchainEpochConfig {
